added rdc file error & started face normal handling
error handling is almost complete, need to remove redundant folders after failed import.
the face and vert normals are being addressed, just adding the correct methodology to provide the best and most consistent results with nanite inside unreal engine.
TODO
Looking into addressing the overlapping UVs that come with joining the models and welding overlapping verts, will likely handle this in a couple of ways using live unwrap and UDIMS if performance issues are required but will likely be solved with the next script when the texture baking and model optimizations have been completed